requirements.txt
langchain
pyPDF2
streamlit
python-dotenv
faiss-cpu
streamlit-extras
tiktoken
openai
langchain-community